fbc: sf.net # 982 - Array descriptors emitted incorrectly in gcc backend
If the symbol is local, then the internally mangled name needs to be
unique so we avoid collisions between structs that have the same name
in different scopes regardless if the types are actually the same type
or not.  And Arrays having descriptors need a unique name for the
descriptor as well since in the gcc backend, we try to type the
descriptor's data pointer to have the element's type (better debug info).

Example of source fixed in this commit:

sub proc1()
  type T
    as integer i
  end type
  static As T A()
end sub

sub proc2()
  type T
    as byte i
  end type
  static As T B()
end sub