fbc: add -gfx command line arg to link against libfbgfx
This already existed as an internal flag, I simply exposed it.  It's useful
under emscripten where we can't write objinfo to wasm (nor js) files. Of course
if you're using emscripten you'd know that you (currently) need libfbgfx.a
specifically, but this is a lot cleaner for portable build systems and
future-proof.