tests: emscripten doesn't support common symbols, so skip COMMON tests
clang (as of v12.0.0) throws an error on the __attribute__((common)) in the
produced .c so these tests prevented building fbc-tests. Emscripten doesn't
support it with -sWASM=0 either, because that still first compiles to wasm .o
files. There's apparently no plan to support common symbols in wasm.

A couple of the tests in namespace/redim-3.bas are kludged to compile but fail.