android: query sysroot in ENABLE_STANDALONE=1
Running a standalone version of fbc to build with
android ndk is not recommended, but if we are,
try and query the sysroot.

If we explicitly passed in -sysroot to fbc, then
go ahead try to query gcc for the needed library
even if we aren't targeting android.

Otherwise, if we are running standalone fbc and not
targeting android and didn't pass in a -sysroot,
just return the computed library file name as was
done previously in fbc.