fbc: string*n intialization
fixed: Copy back string*N passed as argument to zstring parameters

- arguments of STRING*N type are copied to a temporary variable when passed to BYVAL ZSTRING PTR and BYREF ZSTRING parameters (since there is no expectaction that STRING*N will have it's own terminating null character) and copied back after the call
- tests added

fixed: Initialize UDT fields with spaces

changed: Improve on initialization of Types and Unions:

- unions containing STRING*N fields will initialize to spaces if the first field
  is a STRING*N type, otherwise will initialize to zero
- internal: types containing only STRING*N fields or child types containing
  only STRING*N fields will initialize to spaces using a memory fill operation,
  and where other data types are present will induce the creation of an
  implicit constructor

Internal:

Added FB_UDTOPT_HASZEROEDFIELD and FB_UDTOPT_HASFILLEDFIELD to track
fields that can be zeroed or filled during the composition of a UDT.
This allows making decisions about how a field can be initialized.
Arguably not the most efficient method to track this information and
perhaps later we should add to FBS_STRUCT to make decision if it
is more efficient to generate an implicit constructor or inline
assignments even for non-zero and initialized fields.